// src/app/main/account/models/user.model.ts
export class User {
    clientId: number;
    firstName: string;
    lastName: string;
    email: string;
    role: string;
    admin: boolean;
    userType: string;
    clientUserType;
    adminRights: boolean;
    userRole: number;
    clientUserId: number;
    phone: string;
    userTypeDisplayText: string;
}